Rewrite task scheduling
So far git and archive tasks were executed in exactly that order. With this
patch there is no 'type ordering', meaning git and archive tasks can be
interchanged. This opens up dependencies between them which is neither good
nor bad but definitely increases the complexity. Additionally this opens the
possibility to save partial reports in case of a program termination and the
option to continue from that point.