Fix reuse of provided report
Apart from not working correctly it was also implemented in an unsafe way.
It didn't matter which parameter were provided to the tasks - only the version
was checked. Now a hash is generated for the parameter and saved to the report
which has to match if the report should be considered reusable