Improve performance by ~8x using processes
Python has the Global Interpreter Lock (GIL) which makes the usage of multiple
threads kind of pointless if the purpose of them is to speed up the execution.
Threads can be used if they load (large) files from disk or network or for
user interaction. For faster execution processes are more appropriate as they
have their own GIL.

This also makes the FileTasks much nicer and not a pain in the backside